Integrated function with actor
